src/grasp_api/api/v1/provision/base/serializers.py
from pydantic import BaseModel


class ExampleRM(BaseModel):
    pk_example: str
    string_example: str
    string2_example: str


class HealthCheck(BaseModel):
    name: str
    version: str
    description: str
